+++ inst/test/test.R
@@ -276,3 +276,75 @@ p2 <- show_sig_profile(sig2, mode = "SBS", sig_orders = 2)
 
 show_sig_profile_loop(sig2, mode = "SBS")
 show_sig_profile_loop(sig2, mode = "SBS", style = "cosmic", sig_names = c("A", "B", "C"))
+
+
+# Test New CN class -------------------------------------------------------
+
+copyNumberInfo = function(chrom,start,end,CN,CNb=NULL,doVariables=TRUE)
+{
+  # segment lengths
+  print("lengths")
+  lengths = (end-start)/1000000
+  if(!is.null(CNb))
+  {
+    # allelic imbalance
+    print("imbalance")
+    CNa = CN-CNb
+    imba = abs(CNa-CNb)
+    # LOH
+    LOH=pmin(CNa,CNb)
+  } else {
+    imba=NULL
+    LOH=NULL
+  }
+  # combine
+  print("combine")
+  if(!doVariables)
+  {
+    combined = list(CN=CN,
+                    lengths=lengths,
+                    LOH=LOH)
+    return(combined)
+  }
+  LOHstatus = ifelse(LOH==0,"LOH","het")
+  LOHstatus[which(CN==0)] = "homdel"
+  variables = table(paste(cut(CN,
+                              breaks=c(-0.5,1.5,2.5,4.5,Inf),
+                              labels=c("del","neut","dup","amp")),
+                          LOHstatus,
+                          cut(lengths,breaks=c(-0.01,0.01,0.1,1,10,Inf)),
+                          sep=":"))
+  return(variables)
+}
+
+load(system.file("extdata", "toy_segTab.RData",
+                 package = "sigminer", mustWork = TRUE
+))
+set.seed(1234)
+segTabs$minor_cn <- sample(c(0, 1), size = nrow(segTabs), replace = TRUE)
+copyNumberInfo(
+  segTabs$chromosome,
+  segTabs$start,
+  segTabs$end,
+  segTabs$segVal,
+  CNb=segTabs$minor_cn,
+  doVariables=TRUE) -> x
+
+names(x)
+
+x2 <- data.table::fread("inst/extdata/metastatic_tumor.segtab.txt")
+copyNumberInfo(
+  x2$Chromosome,
+  x2$Start.bp,
+  x2$End.bp,
+  x2$modal_cn,
+  x2$subclonal,
+  doVariables=TRUE) -> x2
+
+cn <- read_copynumber(
+  segTabs,
+  seg_cols = c("chromosome", "start", "end", "segVal"),
+  genome_build = "hg19", complement = FALSE, add_loh = TRUE,
+  join_adj_seg = FALSE)
+dt <- cn@data
+dt