+++ tests/testthat/test-roxytest-testexamples-read_copynumber.R
@@ -2,7 +2,7 @@
 
 context("File R/read_copynumber.R: @testexamples")
 
-test_that("Function read_copynumber() @ L75", {
+test_that("Function read_copynumber() @ L78", {
   
   # Load toy dataset of absolute copynumber profile
   load(system.file("extdata", "toy_segTab.RData",
@@ -22,6 +22,8 @@ test_that("Function read_copynumber() @ L75", {
     seg_cols = c("chromosome", "start", "end", "segVal"),
     genome_measure = "wg", complement = TRUE, add_loh = TRUE
   )
+  # Use tally method "S" (Steele et al.)
+  tally_s <- sig_tally(cn, method = "S")
   
   tab_file <- system.file("extdata", "metastatic_tumor.segtab.txt",
     package = "sigminer", mustWork = TRUE
@@ -31,5 +33,6 @@ test_that("Function read_copynumber() @ L75", {
   expect_s4_class(cn, "CopyNumber")
   expect_s4_class(cn_subset, "CopyNumber")
   expect_s4_class(cn2, "CopyNumber")
+  expect_is(tally_s, "list")
 })